Worked on cleaning up app.js, and completed map.js
