+++ miracle-messages/src/Map_Componenets/city_info.js
@@ -3,14 +3,38 @@ import facbook from '../icons/facebook.png';
 import google from '../icons/google.png';
 import gmail from '../icons/gmail.png';
 
-export default class CityInfo extends PureComponent {
+import { getData } from "../Actions/index";
+import { connect } from "react-redux";
+
+
+class CityInfo extends PureComponent {
+
+  constructor(props) {
+        super(props);
+        this.state = {
+          learnMore: false //learn more is a togleinside the pop-ups
+        };
+      }
+
+  componentDidMount() {
+    this.props.getData();
+  }
+
+//this is a toggle that switches the pop-up info upon clicking "learn more" button
+learnMoreToggle = e => {
+  e.preventDefault()
+  console.log(this.props)
+  this.setState({ learnMore: !this.state.learnMore })
+  
+};
+
   render() {
     const { info } = this.props;
 
     let display;
     let button;
 
-    if (this.props.learnMore) {
+    if (this.state.learnMore) {
       display = (
         <div>
           <a
@@ -74,9 +98,104 @@ export default class CityInfo extends PureComponent {
           >
             Join Chapter
           </a>
-          <button onClick={e => this.props.toggle(e)}>{button}</button>
+          <button onClick={e => this.learnMoreToggle(e)}>{button}</button>
         </div>
       </div>
     );
   }
 }
+
+const mapStateToProps = state => {
+  return {
+    chapter_data: state.chapterReducer.chapter_data,
+    fetching: state.chapterReducer.fetching,
+    popupInfo: state.chapterReducer.popupInfo,
+    
+  };
+};
+
+export default connect(
+  mapStateToProps,
+  { getData }
+)(CityInfo);
+
+//*** OLD WORKING COPY OF CITY_INFO PRIOR TO REDUX ***/
+
+// export default class CityInfo extends PureComponent {
+//   render() {
+//     const { info } = this.props;
+
+//     let display;
+//     let button;
+
+//     if (this.props.learnMore) {
+//       display = (
+//         <div>
+//           <a
+//             href="https://www.facebook.com/miraclemessages/"
+//             target="_blank"
+//             rel="noopener noreferrer"
+//           >
+//             Facebook
+//             <span>
+//               <img src={facbook} alt="facbook logo" />
+//             </span>
+//           </a>
+//           <a
+//             href="https://miraclemessages.org"
+//             target="_blank"
+//             rel="noopener noreferrer"
+//           >
+//             Google Hangout
+//             <span>
+//               <img src={google} alt="google logo" />
+//             </span>
+//           </a>
+//           <a
+//             href="https://miraclemessages.org"
+//             target="_blank"
+//             rel="noopener noreferrer"
+//           >
+//             Contact
+//             <span>
+//               <img src={gmail} alt="gmail logo" />
+//             </span>
+//           </a>
+//         </div>
+//       );
+//       button = 'Back';
+//     } else {
+//       display = (
+//         <div>
+//           <p>
+//             Volunteers <span className="number">{info.numvolunteers}</span>
+//           </p>
+//           <p>Delivered Messages</p>
+//           <p>Reunions</p>
+//         </div>
+//       );
+//       button = 'Learn More';
+//     }
+
+//     return (
+//       <div className="popup">
+//         <div className="info">
+//           <h3>{info.location.toUpperCase()}</h3>
+//           {display}
+//         </div>
+
+//         <div className="buttons">
+//           <a
+//             href="https://miraclemessages.org/getinvolved"
+//             target="_blank"
+//             rel="noopener noreferrer"
+//           >
+//             Join Chapter
+//           </a>
+//           <button onClick={e => this.props.toggle(e)}>{button}</button>
+//         </div>
+//       </div>
+//     );
+//   }
+// }
+
