+++ miracle-messages/src/Map_Componenets/map.js
@@ -1,39 +1,124 @@
-import React, { Component } from 'react';
-import MapGL, { Marker, Popup, NavigationControl, StaticMap } from 'react-map-gl';
-import DeckGL from 'deck.gl';
-import CityPin from './city_pin';
-import CityInfo from './city-info';
-import 'mapbox-gl/dist/mapbox-gl.css';
+import React, { Component } from "react";
+import MapGL, { Marker, Popup, NavigationControl} from "react-map-gl";
+import DeckGL from "deck.gl";
+import CityPin from "./city_pin";
+import CityInfo from "./city_info";
+import "mapbox-gl/dist/mapbox-gl.css";
 //import './CSS/MapGl.css';
+import { getData } from "../Actions/index";
+import { connect } from "react-redux";
 
-require('dotenv').config();
+require("dotenv").config();
 
 const TOKEN = process.env.REACT_APP_MAPBOX_TOKEN;
 const STYLE = "mapbox://styles/miraclemessages/cjyhf6b851bii1cq6lr990cf1";
 const WIDTH = "100vw";
 const HEIGHT = "100vh";
 const INITIAL_VIEW_STATE = {
-    latitude: 37.785164,
-    longitude: -100,
-    zoom: 3.5
-}
+  latitude: 37.785164,
+  longitude: -100,
+  zoom: 3.5
+};
+
+
 
 class Map extends Component {
-    // A helper function for moving around the map
-    _updateViewport = viewport => {
-        this.setState({ viewport });
-    };
+  //this fetches the data from the backend
+    componentDidMount() {
+        this.props.getData();
+      }
 
-    render() 
-    {
-        return(
-            <div className="Map">
-                <DeckGL initialViewState={INITIAL_VIEW_STATE} controller={{dragRotate: false}}> 
-                    <StaticMap mapboxApiAccessToken={TOKEN} mapStyle={STYLE}/>
-                </DeckGL>
-            </div>
-        );
-    }
+//this is a toggle that switches the pop-up info upon clicking "learn more" button
+  learnMoreToggle = e => {
+    e.preventDefault();
+    this.setState({ learnMore: !this.state.learnMore });
+  };
+
+
+  _renderCityMarker = (city, index) => {
+    return (
+      <Marker
+        key={`marker-${index}`}
+        latitude={city.latitude}
+        longitude={city.longitude}
+      >
+        <CityPin size={20} onClick={() => {this.setState({ popupInfo: city })}} />
+        {/* <CityPin size={20} /> */}
+      </Marker>
+    );
+  };
+
+  _renderPopup() {
+    if (this.state!=null && this.state.popupInfo){
+      const { popupInfo } = this.state;
+    return (
+      popupInfo && (
+        <Popup
+          className="cityPopup"
+          tipSize={20}
+          offsetTop={-20}
+          latitude={popupInfo.latitude}
+          longitude={popupInfo.longitude}
+          closeButton={true}
+          closeOnClick={false}
+          onClose={() => this.setState({ popupInfo: null })}
+        >
+          <CityInfo
+            info={popupInfo}
+            // toggle={this.learnMoreToggle}
+            learnMore={this.state.learnMore}
+          />
+        </Popup>
+      )
+    );
+  }
 }
 
-export default Map;
+  _updateViewport = viewport => {
+    this.setState({ viewport });
+    console.log(viewport);
+  };
+
+  render() {
+      const { viewport } = this.props;
+    return (
+      <div className="Map">
+        {/* <DeckGL initialViewState={INITIAL_VIEW_STATE} controller={{ dragRotate: false }}>
+          <StaticMap mapboxApiAccessToken={TOKEN} mapStyle={STYLE}>
+            {this.props.chapter_data.map(this._renderCityMarker)}
+          </StaticMap>
+        </DeckGL> */}
+        <MapGL
+          //onClick={() => this.setState({ popupInfo: null })}
+          {...viewport}
+          width="100vw"
+          height="100vh"
+          mapStyle="mapbox://styles/miraclemessages/cjyhf6b851bii1cq6lr990cf1"
+          onViewportChange={this._updateViewport}
+          mapboxApiAccessToken={TOKEN}
+          minZoom={3}
+          maxPitch={0}
+          dragRotate={false}
+        >
+          {this.props.chapter_data.map(this._renderCityMarker)}
+          {this._renderPopup()}
+        </MapGL>
+      </div>
+    );
+  }
+}
+
+const mapStateToProps = state => {
+  return {
+    chapter_data: state.chapterReducer.chapter_data,
+    fetching: state.chapterReducer.fetching,
+    popupInfo: state.chapterReducer.popupInfo,
+    learnMore: state.chapterReducer.learnMore,
+    viewport: state.chapterReducer.viewport
+  };
+};
+
+export default connect(
+  mapStateToProps,
+  { getData }
+)(Map);
