+++ miracle-messages/src/index.js
@@ -5,17 +5,29 @@ import App from './App';
 import * as serviceWorker from './serviceWorker';
 import thunk from 'redux-thunk';
 import logger from 'redux-logger';
-import { createStore, applyMiddleware } from "redux";
+import { createStore, applyMiddleware, compose } from "redux";
 import { Provider } from "react-redux";
 import combineReducers from './Reducers';
 
+const composeEnhancers =
+  typeof window === 'object' &&
+  window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ ?   
+    window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__({
+      // Specify extension’s options like name, actionsBlacklist, actionsCreators, serialize...
+    }) : compose;
 
-
-const store = createStore(
-    // name of reducers 
-    combineReducers,
-    applyMiddleware(thunk, logger)
+const enhancer = composeEnhancers(
+  applyMiddleware(thunk, logger),
+  // other store enhancers if any
 );
+const store = createStore(combineReducers, enhancer);
+
+// const store = createStore(
+//     // name of reducers 
+//     combineReducers,
+//     applyMiddleware(thunk, logger),
+//     window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
+// );
 
 console.log(store)
